+++ src/main/java/pl/put/poznan/transformer/logic/BaseTransformer.java
@@ -1,0 +1,8 @@
+package pl.put.poznan.transformer.logic;
+
+public class BaseTransformer implements TextTransformer {
+    public String transform(String text){
+        logger.debug("BaseTransformer input/output: "+text);
+        return text;
+    }
+}
